refactor(env): extract shared envPositiveInt helper to env-utils
Dedup the byte-divergent envPositiveInt (credential-proxy, LIA-236) and
parsePositiveMsEnv (evolution-client, LIA-235) into a single src/env-utils.ts
helper; consolidate the duplicated unit tests into env-utils.test.ts. The
deliberate dynamic process.env[name] index (flag_lint avoidance) is preserved.

LIA-293
